Grading
See? You can do it!

Luckily in the real world due dates are far more fluid than this course. A mini is maybe just a bit too short.

Only a few points off for not using the group/select to get the distinct values from columns and not having your filter react for your polygons.

It was a pleasure teaching you this mini, good luck with the rest of your final year!